requirements.txt
asyncio-extras==1.3.2
Telethon==1.28.5
cryptg==0.4.0
hachoir==3.1.3
Pillow==10.0.1
pillow_heif==0.21.0